# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HyperMind AI API — multi-agent AI SaaS platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface SuccessMessage {
  message: string;
}

export interface User {
  id: number;
  email: string;
  username: string;
  createdAt: string;
}

export interface UserRegistration {
  email: string;
  /** @minLength 2 */
  username: string;
  /** @minLength 8 */
  password: string;
}

export interface UserLogin {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  /** @nullable */
  agentId?: string | null;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationInput {
  title: string;
  agentId?: string;
}

export interface GeminiMessageInput {
  content: string;
  agentId?: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  /** @nullable */
  agentId?: string | null;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiImageInput {
  prompt: string;
}

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}

export interface GeminiError {
  error: string;
}

export interface Agent {
  id: string;
  name: string;
  description: string;
  icon: string;
  systemPrompt: string;
  capabilities: string[];
  conversationCount: number;
}

export interface MemoryEntry {
  id: number;
  key: string;
  value: string;
  /** @nullable */
  agentId?: string | null;
  createdAt: string;
}

export interface MemoryInput {
  key: string;
  value: string;
  agentId?: string;
}

export interface PromptTemplate {
  id: number;
  name: string;
  content: string;
  /** @nullable */
  agentId?: string | null;
  category: string;
  isBuiltin: boolean;
  createdAt: string;
}

export interface PromptTemplateInput {
  name: string;
  content: string;
  agentId?: string;
  category: string;
}

export interface UploadedFile {
  id: number;
  filename: string;
  originalName: string;
  mimeType: string;
  size: number;
  /** @nullable */
  conversationId?: number | null;
  createdAt: string;
}

export interface FileUploadInput {
  filename: string;
  originalName: string;
  mimeType: string;
  size: number;
  conversationId?: number;
}

export interface DashboardStats {
  totalConversations: number;
  totalMessages: number;
  totalMemories: number;
  totalFiles: number;
  /** @nullable */
  mostUsedAgent?: string | null;
  messagesThisWeek: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  /** @nullable */
  agentId?: string | null;
  createdAt: string;
}

export type SearchMemoryParams = {
  q: string;
};

export type ListTemplatesParams = {
  agentId?: string;
};