# lib/api-zod/src/generated/types/geminiImageOutput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HyperMind AI API — multi-agent AI SaaS platform
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiImageOutput {
  b64_json: string;
  mimeType: string;
}